Split init commands
